+++ web/main.dart
@@ -10,6 +10,7 @@ import 'package:hm_animations/src/services/authentication_service/authentication
 import 'package:hm_animations/src/services/group_service/group_service.dart';
 import 'package:hm_animations/src/services/i18n_service/i18n_service.dart';
 import 'package:hm_animations/src/services/network_service/network_service.dart';
+import 'package:hm_animations/src/services/settings_service/settings_service.dart';
 import 'package:hm_animations/src/services/storage_service/storage_service.dart';
 import 'package:hm_animations/src/services/user_service/user_service.dart';
 
@@ -24,6 +25,7 @@ import 'main.template.dart' as self;
     ClassProvider(NetworkService),
     ClassProvider(UserService),
     ClassProvider(GroupService),
+    ClassProvider(SettingsService),
   ],
 )
 final InjectorFactory injector = self.injector$Injector;